- made WeatherParser class for parsing JSON data
