+++ app/src/main/java/com/geekbrains/anasdroweather2/model/MyData.java
@@ -26,7 +26,7 @@ public class MyData implements Observable {
 
 
 
-    Exception exceptionWhileLoading;
+    private Exception exceptionWhileLoading;
