+++ app/src/main/java/com/geekbrains/anasdroweather2/ui/home/CurrentWeatherFragment.java
@@ -152,7 +152,6 @@ int currentTemp;
 
     @Override
     public void updateInterfaceViewData() {
-
 //
         windTextView.setVisibility(interfaceChanger.getIsWind());
         pressureTextView.setVisibility(interfaceChanger.getIsPressure());
